Refactor where finding the logged-in user happens
